+++ pkg/loadoperations/testdata/operations/level/nested/one.ts
@@ -1,0 +1,12 @@
+import { createOperation, z } from '../../../generated/wundergraph.factory';
+
+export default createOperation.query({
+	input: z.object({
+		a: z.number(),
+	}),
+	handler: async ({ input }) => {
+		return {
+			b: input.a,
+		};
+	},
+});