+++ testapps/mtls/.wundergraph/wundergraph.config.ts
@@ -45,7 +45,4 @@ configureWunderGraphApplication({
 	security: {
 		enableGraphQLEndpoint: true,
 	},
-	dotGraphQLConfig: {
-		hasDotWunderGraphDirectory: false,
-	},
 });